# More optimization (part 1)