--- a/VitalRouter.sln
+++ b/VitalRouter.sln
@@ -16,6 +16,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{CBCBF5DB
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VitalRouter.SourceGenerator.MRuby", "src\VitalRouter.SourceGenerator.MRuby\VitalRouter.SourceGenerator.MRuby.csproj", "{D9C22D5E-6186-481B-AAFE-FE56F92BE75E}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VitalRouter.Benchmark", "tests\VitalRouter.Benchmark\VitalRouter.Benchmark.csproj", "{DB9C34F3-EEB9-49AC-A3C6-DF93BECB034A}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -46,6 +48,10 @@ Global
 		{D9C22D5E-6186-481B-AAFE-FE56F92BE75E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D9C22D5E-6186-481B-AAFE-FE56F92BE75E}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{D9C22D5E-6186-481B-AAFE-FE56F92BE75E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{DB9C34F3-EEB9-49AC-A3C6-DF93BECB034A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{DB9C34F3-EEB9-49AC-A3C6-DF93BECB034A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{DB9C34F3-EEB9-49AC-A3C6-DF93BECB034A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{DB9C34F3-EEB9-49AC-A3C6-DF93BECB034A}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{5EB8B2E1-8C01-4D6F-96B3-F4DC0E8BE4C6} = {40063808-A412-46A8-A4E2-E77256444895}
@@ -54,5 +60,6 @@ Global
 		{E6C26E17-4F85-48C8-B5C9-1DC1549C362D} = {CBCBF5DB-09AA-4FAD-A9D1-C148885BB189}
 		{163E0808-D858-4C31-9C37-A0B50B912608} = {CBCBF5DB-09AA-4FAD-A9D1-C148885BB189}
 		{D9C22D5E-6186-481B-AAFE-FE56F92BE75E} = {40063808-A412-46A8-A4E2-E77256444895}
+		{DB9C34F3-EEB9-49AC-A3C6-DF93BECB034A} = {CBCBF5DB-09AA-4FAD-A9D1-C148885BB189}
 	EndGlobalSection
 EndGlobal