--- a/src/VitalRouter.Unity/Assets/VitalRouter/Runtime/AnonymousSubscriber.cs
+++ b/src/VitalRouter.Unity/Assets/VitalRouter/Runtime/AnonymousSubscriber.cs
@@ -16,7 +16,7 @@ public static class SubscribableAnonymousExtensions
     [Obsolete("Use SubscribeAwait instead")]
     public static Subscription Subscribe<T>(
         this ICommandSubscribable subscribable,
-        Func<T, PublishContext, ValueTask> callback)
+        PublishContinuation<T> callback)
         where T : ICommand
     {
         return subscribable.Subscribe(new AsyncAnonymousSubscriber<T>(callback));
@@ -24,7 +24,7 @@ public static class SubscribableAnonymousExtensions
 
     public static Subscription SubscribeAwait<T>(
         this ICommandSubscribable subscribable,
-        Func<T, PublishContext, ValueTask> callback,
+        PublishContinuation<T> callback,
         CommandOrdering? ordering = null)
         where T : ICommand
     {
@@ -34,10 +34,10 @@ public static class SubscribableAnonymousExtensions
 
 class AsyncAnonymousSubscriber<T> : IAsyncCommandSubscriber where T : ICommand
 {
-    Func<T, PublishContext, ValueTask> callback;
+    readonly PublishContinuation<T> callback;
     readonly ICommandInterceptor? commandOrdering;
 
-    public AsyncAnonymousSubscriber(Func<T, PublishContext, ValueTask> callback, CommandOrdering? ordering = null)
+    public AsyncAnonymousSubscriber(PublishContinuation<T> callback, CommandOrdering? ordering = null)
     {
         this.callback = callback;
         commandOrdering = ordering switch
@@ -70,6 +70,15 @@ class AsyncAnonymousSubscriber<T> : IAsyncCommandSubscriber where T : ICommand
             return callback(commandCasted, context);
         }
         return default;
+    }
+
+    internal ValueTask ReceiveInternalAsync(T command, PublishContext context)
+    {
+        if (commandOrdering != null)
+        {
+            return commandOrdering.InvokeAsync(command, context, callback);
+        }
+        return callback(command, context);
     }
 }