--- a/src/VitalRouter.Unity/Assets/VitalRouter/Runtime/CommandPublisherExtensions.cs.meta
+++ b/src/VitalRouter.Unity/Assets/VitalRouter/Runtime/CommandPublisherExtensions.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 704a7c4f395fc46ebb9da9f4741b9bcd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant: