--- a/src/VitalRouter.Unity/Assets/VitalRouter/Runtime/Internal/ReusableWhenAllSource.cs.meta
+++ b/src/VitalRouter.Unity/Assets/VitalRouter/Runtime/Internal/ReusableWhenAllSource.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1835a63b129084a0b9a67dc151e0d7e5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant: