--- a/src/VitalRouter.Unity/Assets/VitalRouter/Runtime/Internal/WhenAllUtility.cs.meta
+++ b/src/VitalRouter.Unity/Assets/VitalRouter/Runtime/Internal/WhenAllUtility.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 28816ded2b864fb18d90e4e1631c354b
-timeCreated: 1703596210