--- a/tests/VitalRouter.Benchmark/Program.cs
+++ b/tests/VitalRouter.Benchmark/Program.cs
@@ -1,0 +1,18 @@
+﻿using System.Reflection;
+using BenchmarkDotNet.Configs;
+using BenchmarkDotNet.Diagnosers;
+using BenchmarkDotNet.Jobs;
+using BenchmarkDotNet.Running;
+
+BenchmarkSwitcher.FromAssembly(Assembly.GetEntryAssembly()!).Run(args);
+
+class BenchmarkConfig : ManualConfig
+{
+    public BenchmarkConfig()
+    {
+        AddDiagnoser(MemoryDiagnoser.Default);
+        AddJob(Job.ShortRun
+            .WithWarmupCount(5)
+            .WithIterationCount(50));
+    }
+}