The following benchmark was published and we decided to improve our performance on this benchmark.

This change will make it twice as fast for the time being.

NOTE: The benchmark is not available in the
The remaining issue seems to be the presence of PublishContext.
If this is removed, it will probably be close to the same speed as the first place.
However, there are no plans to do so.
Also, VitalRouter focuses on async use cases where there is an interceptor.

## Before (Publish (sync))

```
| Method                              | Mean        | Error    | StdDev    | Gen0   | Allocated |
|------------------------------------ |------------:|---------:|----------:|-------:|----------:|
| 'Publish (ZeroMessenger)'           |    48.21 ns | 0.782 ns |  0.731 ns |      - |         - |
| 'Publish (MessagePipe)'             |    64.46 ns | 1.116 ns |  1.044 ns |      - |         - |
| 'Publish (R3 Subject)'              |    72.90 ns | 1.468 ns |  2.610 ns |      - |         - |
| 'Publish (System.Reactive Subject)' |    80.55 ns | 1.634 ns |  2.181 ns |      - |         - |
| 'Publish (VitalRouter)'             |   213.04 ns | 4.229 ns |  5.928 ns |      - |         - |
| 'Publish (MediatR)'                 |   257.42 ns | 4.083 ns |  3.819 ns |      - |      64 B |
| 'Publish (PubSub)'                  |   524.70 ns | 9.877 ns |  9.239 ns |      - |     248 B |
| 'Publish (Prism)'                   | 1,311.63 ns | 9.929 ns | 25.091 ns | 0.1000 |    1552 B |
```

## After (Publish (sync))

```
| Method                              | Mean        | Error     | StdDev     | Median      | Gen0   | Allocated |
|------------------------------------ |------------:|----------:|-----------:|------------:|-------:|----------:|
| 'Publish (ZeroMessenger)'           |    48.55 ns |  0.626 ns |   0.585 ns |    48.55 ns |      - |         - |
| 'Publish (MessagePipe)'             |    63.77 ns |  0.853 ns |   0.756 ns |    63.69 ns |      - |         - |
| 'Publish (R3 Subject)'              |    71.61 ns |  1.259 ns |   1.499 ns |    71.51 ns |      - |         - |
| 'Publish (System.Reactive Subject)' |    76.56 ns |  0.656 ns |   0.582 ns |    76.32 ns |      - |         - |
| 'Publish (VitalRouter)'             |    83.31 ns |  0.802 ns |   0.750 ns |    83.39 ns |      - |         - |
| 'Publish (MediatR)'                 |   260.63 ns |  2.586 ns |   2.419 ns |   260.59 ns |      - |      64 B |
| 'Publish (PubSub)'                  |   515.71 ns |  6.076 ns |   5.683 ns |   516.50 ns |      - |     248 B |
| 'Publish (Prism)'                   | 1,374.80 ns | 39.718 ns | 103.232 ns | 1,334.21 ns | 0.1000 |    1552 B |
```